Tried in vain to increase the maximum number of ecoregions by changing MapCode from ushort to uint. Archiving the code. Will replace with previous version(s) on next commit.
git-svn-id: http://Marc-PC/svn/Full@3606 85f15e0e-5137-b144-a863-98bd542365a3

+++ model/branches/issue_43/core/src/Implementation/ecoregions/DatasetParser.cs
@@ -33,11 +33,11 @@ namespace Landis.Ecoregions
             IEditableDataset dataset = new EditableDataset();
 
             Dictionary <string, int> nameLineNumbers = new Dictionary<string, int>();
-            Dictionary <ushort, int> mapCodeLineNumbers = new Dictionary<ushort, int>();
+            Dictionary <uint, int> mapCodeLineNumbers = new Dictionary<uint, int>();
 
             InputVar<string> name = new InputVar<string>("Name");
             InputVar<string> description = new InputVar<string>("Description");
-            InputVar<ushort> mapCode = new InputVar<ushort>("Map Code");
+            InputVar<uint> mapCode = new InputVar<uint>("Map Code");
             InputVar<bool> active = new InputVar<bool>("Active");
 
             while (! AtEndOfInput) {
